Merged from dev

+++ rtengine/procparams.h
@@ -531,7 +531,7 @@ struct SharpenMicroParams {
     bool    matrix;
     double  amount;
     double  contrast;
-    double  uniformity;
+    int     uniformity;
 
     SharpenMicroParams();
